requirements.txt
Django==1.11.7
django-ckeditor==5.3.1
django-jet==1.0.6
django-js-asset==0.1.1
django-material==1.1.1
djangorestframework==3.7.3
pytz==2017.3
